+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3](https://github.com/anson-vandoren/pagepouch-rs/compare/v0.1.2...v0.1.3) - 2025-09-12
+
 ### Changed
 
 - Rejiggered the visuals of the login card.